+++ rollup.config.js
@@ -1,7 +1,6 @@
 import resolve from "rollup-plugin-node-resolve";
 import commonjs from "rollup-plugin-commonjs";
 import sourceMaps from "rollup-plugin-sourcemaps";
-import camelCase from "lodash.camelcase";
 import typescript from "rollup-plugin-typescript2";
 import json from "rollup-plugin-json";
 
@@ -9,10 +8,22 @@ const pkg = require("./package.json");
 
 const libraryName = "hybrid-js-utils";
 
+function toCamelCase(s) {
+  return s
+  .replace(/_/g, " ")
+  .replace(/\s(.)/g, ($1) => {
+    return $1.toUpperCase();
+  })
+  .replace(/\s/g, "")
+  .replace(/^(.)/, ($1) => {
+    return $1.toLowerCase();
+  });
+}
+
 export default {
   input: `src/${libraryName}.ts`,
   output: [
-    { file: pkg.main, name: camelCase(libraryName), format: "umd", sourcemap: true },
+    { file: pkg.main, name: toCamelCase(libraryName), format: "umd", sourcemap: true },
     { file: pkg.module, format: "es", sourcemap: true }
   ],
   // Indicate here external modules you don't wanna include in your bundle (i.e.: 'lodash')